Adapt key_costs and unbalancing_positions for moonlander layout.
Increase costs for ring finger on upper row from 9 to 15, as movement seems more exhausting than index finger to upper row (12). Also index is about same length than ring finger, so it should receive a similar value.

As ring and index finger are about the same length, the hand disbalance for these two fingers should be the same. Therefore, increase value for ring finger on upper row from 0 to 0.1 to match the value for index.

Reduce disbalance for lateral movement of index finger to 0.5, as disbalance is not as severe as for pinky (1).